Refactor `state` and `start`
State was moved to a separate file as it became too large. Two associated methods were added to it. `poll` and `handle_events`.

poll is simply polling the events. And handle_events is as the name suggests, handling the events.

Types added:

- `Connection` was added to wrap around TcpStream and the associated ServerConfig.

Types changed:
- `ServerConfig` had some fields that were omitted. These being `all_listeners` (moved into the type `Listener`), and `to_close`. Closing happens in `handle_existing_connection` now.
- `Client` was improperly named. Renamed it to `Listener`.

Also add a runtime check for OS for the set_linger